% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_extent.R
\name{as_double}
\alias{as_double}
\alias{as_double}
\alias{as_double.Extent}
\alias{as_double}
\alias{as_integer}
\alias{as_double}
\alias{as_integer.Extent}
\title{Atomic vector extent}
\usage{
as_double(x, ...)

\method{as_double}{Extent}(x, ...)

as_integer(x, ...)

\method{as_integer}{Extent}(x, ...)
}
\arguments{
\item{x}{a \code{\link[raster]{extent}}}

\item{...}{unused}
}
\value{
numeric vector
}
\description{
Coerce a \code{\link[raster]{extent}} to an atomic vector of \code{c(xmin(x), xmax(x), ymin(x), ymax(x))}.
}
\details{
Note that \code{as_integer} results in truncation, see rasterOps for positive buffering.
}
\seealso{
base::as.double
}
\keyword{internal}
