% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_all_sample_names}
\alias{get_all_sample_names}
\title{Extract sample names from metadata}
\usage{
get_all_sample_names(metadata)
}
\arguments{
\item{metadata}{A `dataframe` or `tibble` containing the metadata.}
}
\value{
vector of sample names
}
\description{
Extract sample names from metadata
}
\examples{
get_all_sample_names(metadata)
}
