% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_processing.R
\name{get_condition_summary}
\alias{get_condition_summary}
\title{Summary of metadata conditions}
\usage{
get_condition_summary(metadata)
}
\arguments{
\item{metadata}{dataframe or tibble of metadata}
}
\value{
list of tibbles each containing a condition type defined in the metadata,
and the number of instances of each condition.
}
\description{
Summary of metadata conditions
}
\examples{
get_condition_summary(metadata)
}
