% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spex-package.R
\docType{data}
\name{lux}
\alias{lux}
\title{The 'lux' Spatial Polygons from the 'raster' package.}
\format{
\code{\link[sp]{SpatialPolygonsDataFrame}} with columns:
\itemize{
\item ID_1
\item NAME_1
\item ID_2
\item NAME_2
\item AREA
}
}
\description{
The 'lux' Spatial Polygons from the 'raster' package.
}
\examples{
library(sp)
plot(lux)
}
