% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spex.R
\name{parse_leaf_extent}
\alias{parse_leaf_extent}
\title{{"_southWest":{"lat":-1.307259612275665,"lng":23.411865234375},"_northEast":{"lat":6.937332868878443,"lng":31.904296875000004}}'}
\usage{
parse_leaf_extent(x)
}
\description{
{"_southWest":{"lat":-1.307259612275665,"lng":23.411865234375},"_northEast":{"lat":6.937332868878443,"lng":31.904296875000004}}'
}
