% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{dt_setup}
\alias{dt_setup}
\title{Title}
\usage{
dt_setup(
  data,
  n_rows = 10,
  lineHeight = "80\%",
  dom_opt = "tip",
  cols_to_round = NULL,
  dt_options = NULL,
  sig_digits = 3,
  regex = FALSE,
  selection = "single",
  table_lengths = c(10, 20, 50, 100),
  filter_pos = "none",
  show_rownames = FALSE
)
}
\arguments{
\item{data}{tibble or dataframe}

\item{n_rows}{number of rows to display (integer)}

\item{lineHeight}{height of each row in table (percentage)}

\item{dom_opt}{dom options for table components, single character string. l-length changing input control, f-filtering input, t-table, i-table info, p-pagination, they will be displayed in the order supplied. 
Default: "tip" (table, info, pagination)}

\item{cols_to_round}{columns to round (vector of column numbers). 
See sig_digits.  Default: NULL}

\item{dt_options}{list of options to supply to datatable() Default: NULL, options
are taken from the relevant supplied parameters i.e. dom, pageLength = n_rows, 
lengthMenu = table_lengths. This argument is for if further customisation is required.}

\item{sig_digits}{Number of significant digits to round the columns supplied in 
cols_to_round to default: 3}

\item{regex}{default: false. If TRUE, allows regex searchng of the table e.g. gene1 | gene2}

\item{selection}{row selection, one of "multiple", "single", default: single}

\item{table_lengths}{integer vector of available table lengths. Default: c(10,20,50,100).
This will only show if dom_opt includes "l"}

\item{filter_pos}{One of c("none", "bottom", "top"), where to place column filters. Default: "none"}

\item{show_rownames}{boolean whether to show rownames or not. Default: FALSE}
}
\value{
DT::datatable() object
}
\description{
Title
}
\examples{
dt_setup(iris)
}
