% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitude.R
\name{latitudecircle}
\alias{latitudecircle}
\title{Latitude circle}
\usage{
latitudecircle(latitude = 0,
  proj = "+proj=stere +lon_0=0 +lat_0=-90 +lat_ts=-71 +ellps=WGS84",
  lonlim = c(-180, 180))
}
\arguments{
\item{latitude}{latitude value for the boundary, defaults to 0}

\item{proj}{map projection to use, defaults to southern Polar Stereographic true scale at -71S}

\item{lonlim}{the range of longitude to use, defaults to entire globe}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Create a circular polygon using a latitude value in a map projection. The longitude range can be
modified from global to give a portion of a circle.
}
