% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitude.R
\name{latmask}
\alias{latmask}
\title{Latitude mask for polar raster}
\usage{
latmask(x, latitude = 0, southern = TRUE, trim = FALSE, ...)
}
\arguments{
\item{x}{a raster layer}

\item{latitude}{maximum latitude  (effectively a minimum latitude if \code{southern = FALSe})}

\item{southern}{flag for whether south-polar context is used, default is \code{TRUE}}

\item{trim}{if \code{TRUE} runs \code{raster::trim} on the result, to remove \code{NA} margin}

\item{...}{ignored currently}
}
\value{
RasterLayer
}
\description{
Latitude mask for polar raster
}
\examples{
 
library(raster)
plot(latmask(ice, -60))
plot(latmask(ice, -60, trim = TRUE))
}
