% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spex.R
\name{spex}
\alias{spex}
\alias{spex.default}
\alias{spex.Extent}
\alias{spex.sf}
\title{Polygon extent}
\usage{
spex(x, crs, byid = FALSE, .id, ..., clipboard = FALSE)

\method{spex}{default}(x, crs, byid = FALSE, .id, ...,
  clipboard = FALSE)

\method{spex}{Extent}(x, crs, byid = FALSE, .id, ...,
  clipboard = FALSE)

\method{spex}{sf}(x, crs, byid = FALSE, .id, ..., clipboard = FALSE)
}
\arguments{
\item{x}{any object with a \code{\link[raster]{Extent}}}

\item{crs}{a projection string}

\item{byid}{return a separate object for every input sub-object (not yet implemented)}

\item{.id}{optional name for output attribute name}

\item{...}{arguments for methods}

\item{clipboard}{WIP this special-case allows x to be the result of the leafem clipboard copy process}
}
\value{
'SpatialPolygonsDataFrame'
}
\description{
Create Spatial Polygons with projection metadata from a 'Spatial Extent'.
}
\section{Warning}{
 Please note that an extent converted to polygons consists
of only four unique coordinates, and so this is not necessarily suited for
projection transformations.
}

\examples{
library(raster)
data(lux)
exlux <- spex(lux)

plot(lux)
plot(exlux, add = TRUE)

## put an extent and a CRS together
spex(extent(0, 1, 0, 1), crs = "+proj=laea +ellps=WGS84")
}
\seealso{
This pattern is displayed in the example code for \code{\link[raster]{cover}}.
}
