% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spex.R
\name{xlim}
\alias{xlim}
\title{Axis ranges from extent}
\usage{
xlim(x, ...)
}
\arguments{
\item{x}{any object with an extent understood by `spex`}

\item{...}{reserved for future methods}
}
\description{
Functions `xlim` and `ylim` return the two-value counterparts of an extent.
}
\details{
Any projection metadata is dropped since this is a one-dimensional entity.
}
